utilities/hcm-custom-reports/requirements.txt
openpyxl>=3.1.2
pandas>=2.0.3
numpy>=1.26.2
python-dateutil>=2.8.2
pytz>=2023.3
xlsxwriter>=3.1.2
xlrd>=2.0.1
requests>=2.31.0
python-decouple>=3.8